fix: V1.4 Now compiling on Linux + brought over asset renaming

+++ Engine/src/Graphics/Camera.cpp
@@ -11,7 +11,7 @@ namespace IonixEngine
     {
     }
 
-    void Camera::SetZoom(SDL_Renderer* renderer, int zoom)
+    void Camera::SetZoom(SDL_Renderer* renderer, float zoom)
     {
         if (zoom < 0.2f) zoom = 0.2f;
         if (zoom > 5.0f) zoom = 5.0f;
